scene functions

+++ types.h
@@ -8,7 +8,7 @@
 #define global_variable static
 
 typedef float f32;
-typedef int s32;
+typedef double f64;
 typedef unsigned int u32;
 typedef unsigned char u8;
 typedef int bool32;
